Updated test suite with extra z3 tests for coverage
