+++ tests/test_cfg/test_cfg_generator.py
@@ -39,28 +39,6 @@ def create_cfg():
 
 
 @pytest.fixture
-def create_cfg_z3():
-    """Create the CFG for each test that uses the code in a file fixture."""
-    # Setup: store the paths of the files being used/created
-    script_name = "file_fixtures/my_file.py"
-    dot_file_path = os.path.splitext(os.path.basename(script_name))[0] + ".gv"
-    svg_file_path = dot_file_path[:-3] + ".svg"
-
-    # Create the graphviz files using my_file.py
-    cfg_generator.generate_cfg(mod=script_name, auto_open=False, z3_enabled=True)
-
-    # Open the actual graphviz file for reading
-    gv_file_io = open(dot_file_path)
-
-    yield dot_file_path, svg_file_path, gv_file_io
-
-    # Teardown: close any open file IO and remove the graphviz generated files
-    gv_file_io.close()
-    os.remove(dot_file_path)
-    os.remove(svg_file_path)
-
-
-@pytest.fixture
 def create_cfg_funcs_only():
     """Create a CFG for each test that uses the code in a file fixture.
 
@@ -174,23 +152,6 @@ def test_z3_dependency_uninstalled(snapshot, create_cfg) -> None:
     # Check that the contents match with the snapshot
     snapshot.snapshot_dir = "snapshots"
     snapshot.assert_match(gv_file_io.read(), "my_file.gv")
-
-
-def test_cfg_z3_enabled(snapshot, create_cfg_z3) -> None:
-    """Test verifies that `generate_cfg` correctly creates a graphviz dot file with the expected content when z3
-    enabled."""
-    _, _, gv_file_io = create_cfg_z3
-
-    # Check that the contents match with the snapshot
-    snapshot.snapshot_dir = "snapshots"
-    snapshot.assert_match(gv_file_io.read(), "my_file.gv")
-
-
-@patch.dict("sys.modules", {"python_ta.transforms.z3_visitor": None})
-def test_cfg_z3_failed_import(caplog) -> None:
-    """Test verifies that `generate_cfg` handles ImportError appropriately."""
-    with pytest.raises(ImportError):
-        cfg_generator.generate_cfg(z3_enabled=True)
 
 
 if __name__ == "__main__":
